requirements.txt
twilio==9.3.2
python-dotenv==1.0.1
schedule==1.2.2